requirements.txt
# Video Processing
opencv-python==4.8.1.78
moviepy>=1.0.3
yt-dlp>=2025.3.31

# Text Processing
pytesseract==0.3.10
langdetect>=1.0.9
nltk>=3.6.0
scikit-learn>=0.24.0

# Document Processing
python-docx==1.0.1
Pillow>=8.0.0

# Utilities
tqdm>=4.65.0
pyyaml==6.0.1
PyQt6==6.6.1

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0 